some more restructuring
