--- comms/trending/api/playlists.go
@@ -1,1 +1,0 @@
-package api
