--- comms/trending/api/tracks.go
@@ -1,1 +1,0 @@
-package api
