+++ comms/trending/handlers/playlists.go
@@ -1,0 +1,1 @@
+package handlers
