+++ comms/trending/handlers/tracks.go
@@ -1,0 +1,3 @@
+package handlers
+
+func GetTrendingTracksHandler()
